mudanças no animal, criação das multas mas deu problema e eu nao to afim de mexer agora

+++ src/main/java/org/projetoverdinho/projetoverdinhoapi/repository/AnimalRepository.java
@@ -3,5 +3,5 @@ package org.projetoverdinho.projetoverdinhoapi.repository;
 import org.projetoverdinho.projetoverdinhoapi.entity.AnimalModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AnimalRepository extends JpaRepository<AnimalModel, Integer> {
+public interface AnimalRepository extends JpaRepository<AnimalModel, Long> {
 }
